suggestion rework
