+++ react-training/src/day5/GraphQL/client.js
@@ -7,20 +7,6 @@ import {
     uri: "https://countries.trevorblades.com/graphql",
     cache: new InMemoryCache(),
   });
-  // const client = ...
   
-  // client
-  //   .query({
-  //     query: gql`
-  //       query GetLocations {
-  //         locations {
-  //           id
-  //           name
-  //           description
-  //         }
-  //       }
-  //     `,
-  //   })
-  //   .then((result) => console.log(result));
     export default client;
